fix(revset): support builtin immutable-head aliases
