+++ shell/mac/Sources/JayJay/Detail/EvologDisplay.swift
@@ -59,9 +59,7 @@ extension EvologEntry {
             changeId: changeId,
             commitId: commitId,
             description: description,
-            author: "",
-            email: "",
-            timestampMillis: timestampMillis,
+            author: .empty(timestampMillis: timestampMillis),
             parents: [],
             bookmarks: [],
             isWorkingCopy: false,
