+++ shell/mac/Sources/JayJay/Shared/CommitAuthor+Helpers.swift
@@ -1,0 +1,7 @@
+import JayJayCore
+
+extension CommitAuthor {
+    static func empty(timestampMillis: Int64 = 0) -> CommitAuthor {
+        CommitAuthor(name: "", email: "", timestampMillis: timestampMillis)
+    }
+}
