+++ shell/mac/Tests/JayJayTests/CommitAuthorTestHelpers.swift
@@ -1,0 +1,5 @@
+import JayJayCore
+
+extension CommitAuthor {
+    static let tester = CommitAuthor(name: "Tester", email: "tester@example.com", timestampMillis: 0)
+}
