+++ shell/mac/Tests/JayJayTests/DAGRebaseGesturePolicyTests.swift
@@ -167,9 +167,7 @@ final class DAGRebaseGesturePolicyTests: XCTestCase {
                 changeId: changeId,
                 commitId: commitId,
                 description: description,
-                author: "Tester",
-                email: "tester@example.com",
-                timestampMillis: 0,
+                author: .tester,
                 parents: parents,
                 bookmarks: bookmarks,
                 isWorkingCopy: false,
