+++ shell/mac/Tests/JayJayTests/DAGRowViewModelTests.swift
@@ -129,9 +129,7 @@ final class DAGRowViewModelTests: XCTestCase {
                 changeId: changeId,
                 commitId: commitId,
                 description: description,
-                author: "Tester",
-                email: "tester@example.com",
-                timestampMillis: 0,
+                author: .tester,
                 parents: [],
                 bookmarks: [],
                 isWorkingCopy: false,
