+++ shell/mac/Tests/JayJayTests/DAGViewModelTests.swift
@@ -89,9 +89,7 @@ final class DAGViewModelTests: XCTestCase {
                 changeId: changeId,
                 commitId: commitId,
                 description: "entry",
-                author: "Tester",
-                email: "tester@example.com",
-                timestampMillis: 0,
+                author: .tester,
                 parents: [],
                 bookmarks: [],
                 isWorkingCopy: false,
